+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @flowglad/server
 
+## 0.9.1
+
+### Patch Changes
+
+- Fix checkout session create
+- Updated dependencies
+  - @flowglad/shared@0.9.1
+
 ## 0.9.0
 
 ### Minor Changes